backend/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=1.8.0
mysql-connector-python>=8.0.0
email-validator>=1.1.3
aiohttp>=3.8.0
python-jose[cryptography]>=3.3.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
fastapi-limiter>=0.1.5